extracted types and split sync/async API in Swift;
