+++ bindings/ergo-lib-ios/Sources/ErgoLib/NodeInfo.swift
@@ -1,0 +1,14 @@
+import Foundation
+import ErgoLibC
+
+class NodeInfo {
+    internal var pointer: NodeInfoPtr
+
+    internal init(withRawPointer ptr: NodeInfoPtr) {
+        self.pointer = ptr
+    }
+
+    deinit {
+        ergo_lib_node_info_delete(self.pointer)
+    }
+}
